requirements.txt
numpy 
pandas 
streamlit
seaborn 
altair
matplotlib.pyplot
